plugins/compass-api/pyproject.toml
[project]
name = "compasslabs_api_plugin_gamesdk"
version = "0.1.0"
authors = [
  { name = "Ben Hack", email = "contact@compasslabs.ai" },
]
description = "CompassLabs Defi API plugin for Python SDK for GAME by Virtuals"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "compass.virtuals_sdk"
]
